feat(cc): Provider implements StatuslineInstaller interface
